path: Move old path implementation to 'legacy' folder.
