+++ libopenage/pathfinding/legacy/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_sources(libopenage
+	a_star.cpp
+	heuristics.cpp
+	path.cpp
+	tests.cpp
+)
