+++ libopenage/pathfinding/legacy/a_star.h
