+++ libopenage/pathfinding/legacy/heuristics.cpp
