+++ libopenage/pathfinding/legacy/path_utils.h
